# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Conduit — LLM Gateway Portal API",
    "version": "0.1.0",
    "description": "API contract cho backend-gateway (Spring Boot), dựng từ docs/erd.dbml. Phạm vi Must-have là chính; endpoint đánh dấu 'Should-have' (RAG, VNPay) định nghĩa trước cho khớp kiến trúc nhưng chưa bắt buộc triển khai ngay (CLAUDE.md mục 6). Sai số cho phép khi triển khai thực tế: +-3 endpoint so với bản này (thêm/bớt nhỏ khi code, không đổi field đã thống nhất mà không báo lại)."
  },
  "servers": [
    { "url": "http://localhost:8081/api", "description": "backend-gateway local" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true }
        }
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 0 },
          "size": { "type": "integer", "example": 20 },
          "total_elements": { "type": "integer", "example": 137 },
          "total_pages": { "type": "integer", "example": 7 }
        }
      },
...
      "AgentType": { "type": "string", "enum": ["text"] },
      "AgentStatus": { "type": "string", "enum": ["draft", "pending", "approved", "rejected", "published", "unpublished"] },
      "Agent": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "creator_id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "introduction": { "type": "string", "nullable": true },
          "agent_type": { "$ref": "#/components/schemas/AgentType" },
          "price_vnd": { "type": "integer", "minimum": 0 },
          "default_credit_granted": { "type": "integer", "minimum": 0 },
          "status": { "$ref": "#/components/schemas/AgentStatus" },
          "reject_reason": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "PaymentMethod": { "type": "string", "enum": ["vnpay", "mock"] },
      "PaymentStatus": { "type": "string", "enum": ["pending", "paid", "failed", "refunded"] },
      "CreatePurchaseRequest": {
        "type": "object",
        "required": ["payment_method"],
        "properties": {
          "payment_method": { "$ref": "#/components/schemas/PaymentMethod" }
        }
      },
      "AgentPurchase": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "format": "uuid" },
          "agent_id": { "type": "string", "format": "uuid" },
          "amount_vnd": { "type": "integer" },
          "default_credit_granted": { "type": "integer" },
          "payment_method": { "$ref": "#/components/schemas/PaymentMethod" },
          "payment_status": { "$ref": "#/components/schemas/PaymentStatus" },
          "transaction_ref": { "type": "string", "description": "unique, dung cho idempotency" },
          "paid_at": { "type": "string", "format": "date-time", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreatePurchaseResponse": {
        "type": "object",
        "properties": {
          "purchase": { "$ref": "#/components/schemas/AgentPurchase" },
          "redirect_url": { "type": "string", "nullable": true, "description": "URL redirect toi VNPay sandbox; null khi payment_method=mock (client tu goi webhook mock de mo phong)." }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": { "description": "Thieu hoac sai JWT", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "NotFound": { "description": "Khong tim thay", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    }
  },
  "paths": {
...
    "/agents/{agentId}/purchases": {
      "post": {
        "tags": ["Purchases & Payment"],
        "summary": "Khoi tao mua agent (tao agent_purchases voi transaction_ref moi, payment_status=pending)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "agentId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePurchaseRequest" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePurchaseResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "description": "Da mua agent nay roi", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/purchases": {
      "get": {
        "tags": ["Purchases & Payment"],
        "summary": "Thu vien cua toi: cac agent user da mua (payment_status = paid), moi nhat truoc",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "type": "object", "properties": { "agent": { "$ref": "#/components/schemas/Agent" }, "purchase": { "$ref": "#/components/schemas/AgentPurchase" } } } }, "page": { "$ref": "#/components/schemas/PageMeta" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/purchases/{transactionRef}": {
      "get": {
        "tags": ["Purchases & Payment"],
        "summary": "Trang thai 1 giao dich mua agent (client poll sau khi redirect ve tu payment)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "transactionRef", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentPurchase" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  }
}